Revert "Implement Fido2 signup on the authentication widget (#61)"
This reverts commit 938772ae60045d25ffc86cf746dfe31dfc0b4ff3.
