+++ package.json
@@ -19,42 +19,42 @@
     },
     "dependencies": {
         "@hypnosphi/recompose": "0.30.5",
-        "@reachfive/identity-core": "1.17.1",
+        "@reachfive/identity-core": "1.15.0",
         "@reachfive/zxcvbn": "1.0.0-alpha.2",
         "char-info": "0.3.2",
         "classnames": "2.2.6",
-        "libphonenumber-js": "1.8.1",
+        "libphonenumber-js": "1.7.54",
         "lodash-es": "4.17.15",
-        "polished": "3.6.7",
+        "polished": "3.6.5",
         "prop-types": "15.7.2",
         "react": "16.13.1",
         "react-dom": "16.13.1",
         "react-transition-group": "4.4.1",
-        "regenerator-runtime": "0.13.7",
+        "regenerator-runtime": "0.13.5",
         "remarkable": "2.0.1",
-        "styled-components": "5.2.0",
-        "validator": "13.1.17"
+        "styled-components": "5.1.1",
+        "validator": "13.1.1"
     },
     "devDependencies": {
-        "@babel/core": "7.11.6",
+        "@babel/core": "7.10.4",
         "@babel/plugin-proposal-class-properties": "7.10.4",
-        "@babel/preset-env": "7.11.5",
+        "@babel/preset-env": "7.10.4",
         "@babel/preset-react": "7.10.4",
-        "@rollup/plugin-commonjs": "15.1.0",
-        "@rollup/plugin-node-resolve": "9.0.0",
+        "@rollup/plugin-commonjs": "13.0.0",
+        "@rollup/plugin-node-resolve": "8.1.0",
         "@rollup/plugin-replace": "2.3.3",
         "@rollup/plugin-url": "5.0.1",
         "@svgr/rollup": "5.4.0",
         "babel-plugin-dynamic-import-node": "2.3.3",
-        "babel-preset-jest": "26.3.0",
+        "babel-preset-jest": "26.1.0",
         "cheerio": "1.0.0-rc.3",
         "core-js": "3.6.5",
         "enzyme": "3.11.0",
-        "enzyme-adapter-react-16": "1.15.5",
-        "jest": "26.4.2",
-        "jest-styled-components": "7.0.3",
+        "enzyme-adapter-react-16": "1.15.2",
+        "jest": "26.1.0",
+        "jest-styled-components": "7.0.2",
         "react-test-renderer": "16.13.1",
-        "rollup": "2.28.2",
+        "rollup": "2.20.0",
         "rollup-plugin-babel": "4.4.0",
         "rollup-plugin-node-builtins": "2.1.2",
         "rollup-plugin-node-globals": "1.4.0",
