+++ src/components/form/formFieldFactory.js
@@ -95,12 +95,6 @@ const predefinedFields = {
         key: 'address.country',
         label: 'address.country',
         ...cfg
-    }),
-    deviceFriendlyName: cfg => simpleField({
-        key: 'friendlyName',
-        label: 'webauthn.friendly.name',
-        required: false,
-        ...cfg
     })
 };
