+++ src/widgets/auth/views/loginWithWebAuthnViewComponent.jsx
@@ -4,7 +4,7 @@ import { email } from '../../../core/validation';
 import { Heading, Link, Alternative, Separator } from '../../../components/miscComponent';
 
 import SocialButtons from '../../../components/form/socialButtonsComponent';
-import { WebAuthnLoginViewButtons } from './../../../components/form/webAuthAndPasswordButtonsComponent';
+import { WebAuthnViewPrimaryButtons } from './../../../components/form/webAuthButtonsComponent';
 import { createForm } from '../../../components/form/formComponent';
 import { simpleField } from '../../../components/form/fields/simpleField';
 
@@ -46,10 +46,10 @@ export default class LoginWithWebAuthnView extends React.Component {
 
         const defaultEmail = session.lastLoginType === 'password' ? session.email : null;
 
-        const webAuthnButtons = (disabled, handleClick) => <WebAuthnLoginViewButtons
+        const webAuthnButtons = (disabled, handleClick) => <WebAuthnViewPrimaryButtons
             disabled={disabled}
             i18n={i18n}
-            onPasswordClick={handleClick} />
+            onClick={handleClick} />
 
         return (
             <div>
